Tables and records removed from academic tracking system; SQL files, data records, and frontend and backend components deleted. Tables include course, department, instructor, programme, session, student, takes, teaches, user, and administrator. Data records for these tables have also been removed. Frontend components include AdminSignIn, SignInSide, and SignUpSide, while backend components include various CRUD operations for the aforementioned tables. SQL files include schemas for each table and records for inserting data into those tables.

+++ backend/controllers/courses.js
@@ -15,10 +15,9 @@ coursesRouter.get("/:id", async (request, response) => {
         return response.status(400).json({ error: "'id' must be a number." });
     }
 
-    const courseList = await dbConn.query(
-        "SELECT id, title, code, dept_name FROM course WHERE id=?",
-        [id]
-    );
+    const courseList = await dbConn.query("SELECT * FROM course WHERE id=?", [
+        id,
+    ]);
 
     if (courseList.length === 0) {
         return response
@@ -39,7 +38,7 @@ coursesRouter.get("/:id/sessions", async (request, response) => {
     }
 
     const courseList = await dbConn.query(
-        "SELECT id, title, code, dept_name FROM course WHERE id=?",
+        "SELECT id, title, code, department_id FROM course WHERE id=?",
         [id]
     );
 
@@ -52,11 +51,13 @@ coursesRouter.get("/:id/sessions", async (request, response) => {
     const [course] = courseList;
 
     const query = `SELECT session_id, start_year, season, 
-    instructor_id, name instructor_name, designation instructor_designation, instructor.dept_name instructor_dept_name 
+    instructor_id, name instructor_name, designation instructor_designation, 
+    instructor.department_id instructor_dept_id
     FROM teaches 
     JOIN session ON session.id = session_id 
     JOIN course ON course_id = course.id 
-    JOIN instructor ON instructor_id = instructor.id WHERE course_id=?;`;
+    JOIN instructor ON instructor_id = instructor.id 
+    WHERE course_id=?;`;
 
     const sessions = await dbConn.query(query, [id]);
 
@@ -82,10 +83,9 @@ coursesRouter.get("/:id/sessions/:session_id", async (request, response) => {
             .json({ error: "'session_id' must be a number." });
     }
 
-    const courseList = await dbConn.query(
-        "SELECT id, title, code, dept_name FROM course WHERE id=?",
-        [id]
-    );
+    const courseList = await dbConn.query("SELECT * FROM course WHERE id=?", [
+        id,
+    ]);
 
     if (courseList.length === 0) {
         return response
@@ -111,11 +111,11 @@ coursesRouter.get("/:id/sessions/:session_id", async (request, response) => {
     const instructorQuery = `SELECT instructor_id, 
     name instructor_name, 
     designation instructor_designation, 
-    instructor.dept_name instructor_dept_name 
+    instructor.department_id instructor_dept_id 
     FROM teaches 
     JOIN session ON session.id = session_id 
     JOIN course ON course_id = course.id 
-    JOIN instructor ON instructor_id = instructor.id 
+    JOIN instructor ON instructor_id = instructor.id
     WHERE course_id=? AND session_id=?;`;
 
     const instructors = await dbConn.query(instructorQuery, [id, session_id]);
@@ -128,16 +128,16 @@ coursesRouter.get("/:id/sessions/:session_id", async (request, response) => {
 
     const studentsQuery = `SELECT student.roll, student.name, student.email, 
     student.programme_id, programme.degree, programme.name programme_name, 
-    programme.dept_name, taught_by instructor_id, 
-    instructor.name instructor_name, 
+    programme.department_id, taught_by instructor_id, 
+    instructor.name instructor_name,
     instructor.designation instructor_designation, 
-    instructor.dept_name instructor_dept_name
+    instructor.department_id instructor_dept_id
     FROM takes 
     JOIN student ON student.roll = student_roll 
     JOIN instructor ON taught_by = instructor.id 
     JOIN course ON course_id = course.id
     JOIN session ON session_id = session.id
-    JOIN programme ON student.programme_id = programme.id
+    JOIN programme ON student.programme_id = programme.id 
     WHERE course_id=? AND session_id=?;`;
 
     const students = await dbConn.query(studentsQuery, [id, session_id]);
@@ -153,7 +153,7 @@ coursesRouter.post("/", async (request, response) => {
         return response.status(403).end();
     }
 
-    let { dept_name, code, title } = request.body;
+    let { department_id, code, title } = request.body;
 
     if (!code) {
         return response.status(400).json({
@@ -167,15 +167,14 @@ coursesRouter.post("/", async (request, response) => {
         });
     }
 
-    if (!dept_name) {
+    if (!department_id) {
         return response.status(400).json({
-            error: "dept_name missing in request body",
+            error: "department_id missing in request body",
         });
     }
 
     code = code.trim();
     title = title.trim();
-    dept_name = dept_name.trim();
 
     if (!title || title.length < 2) {
         return response.status(400).json({
@@ -189,25 +188,25 @@ coursesRouter.post("/", async (request, response) => {
         });
     }
 
-    if (!dept_name || dept_name.length < 2) {
+    if (!department_id instanceof Number || department_id % 1 !== 0) {
         return response.status(400).json({
-            error: "dept_name cannot be less than 2 characters long",
+            error: "department_id has to be a number",
         });
     }
 
     const department = await dbConn.query(
-        "SELECT * FROM department WHERE name=?",
-        [dept_name]
+        "SELECT * FROM department WHERE id=?",
+        [department_id]
     );
     if (department.length === 0) {
         return response.status(400).json({
-            error: "Invalid dept_name",
+            error: "Invalid department_id",
         });
     }
 
     const res = await dbConn.query(
-        "INSERT INTO course (title, code, dept_name) VALUES (?, ?, ?)",
-        [title, code, dept_name]
+        "INSERT INTO course (title, code, department_id) VALUES (?, ?, ?)",
+        [title, code, department_id]
     );
     return response.status(201).json({
         id: res.insertId,
@@ -220,7 +219,7 @@ coursesRouter.put("/:id", async (request, response) => {
     }
 
     const { id } = request.params;
-    let { title, code, dept_name } = request.body;
+    let { title, code, department_id } = request.body;
 
     if (!code) {
         return response.status(400).json({
@@ -234,15 +233,14 @@ coursesRouter.put("/:id", async (request, response) => {
         });
     }
 
-    if (!dept_name) {
+    if (!department_id) {
         return response.status(400).json({
-            error: "dept_name missing in request body",
+            error: "department_id missing in request body",
         });
     }
 
     code = code.trim();
     title = title.trim();
-    dept_name = dept_name.trim();
 
     if (!title || title.length < 2) {
         return response.status(400).json({
@@ -256,25 +254,25 @@ coursesRouter.put("/:id", async (request, response) => {
         });
     }
 
-    if (!dept_name || dept_name.length < 2) {
+    if (!department_id instanceof Number || department_id % 1 !== 0) {
         return response.status(400).json({
-            error: "dept_name cannot be less than 2 characters long",
+            error: "department_id has to be a number",
         });
     }
 
     const department = await dbConn.query(
-        "SELECT * FROM department WHERE name=?",
-        [dept_name]
+        "SELECT * FROM department WHERE id=?",
+        [department_id]
     );
     if (department.length === 0) {
         return response.status(400).json({
-            error: "Invalid dept_name",
+            error: "Invalid department_id",
         });
     }
 
     await dbConn.query(
-        "UPDATE course SET title=?, code=?, dept_name=? WHERE id=?",
-        [title, code, dept_name, id]
+        "UPDATE course SET title=?, code=?, department_id=? WHERE id=?",
+        [title, code, department_id, id]
     );
     return response.status(200).end();
 });
@@ -283,8 +281,15 @@ coursesRouter.delete("/:id", async (request, response) => {
     if (!request.administrator) {
         return response.status(403).end();
     }
-
     const { id } = request.params;
+
+    const courseWithId = await dbConn.query("SELECT * FROM course WHERE id=?", [
+        id,
+    ]);
+    if (courseWithId.length == 0) {
+        return response.status(404).json({ error: "No such course" }).end();
+    }
+
     await dbConn.query("DELETE FROM course WHERE id=?", [id]);
     return response.status(204).end();
 });
